requirements.txt
absl-py==0.13.0
accelerate==0.20.3
aggdraw==1.3.12
ahocorapy==1.6.1
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
antlr4-python3-runtime==4.8
anyio==3.7.1
argh==0.26.2
asn1crypto==1.3.0
astroid==2.3.3
astropy==4.0
atomicwrites==1.3.0
attrs==21.2.0
audioread==3.0.1
autopep8==1.4.4
awesome-align==0.1.6
Babel==2.8.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.2
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.0
blis==0.7.5
bokeh==1.4.0
boto==2.49.0
boto3==1.18.58
botocore==1.21.38
Bottleneck==1.3.2
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
cloudpickle==1.3.0
clyent==1.2.2
colorama==0.4.3
conda==4.8.2
conda-build==3.18.11
conda-package-handling==1.6.0
conda-verify==3.4.2
contextlib2==0.6.0.post1
cryptography==3.4.7
cycler==0.10.0
cymem==2.0.6
Cython==0.29.24
cytoolz==0.10.1
dask==2.11.0
dataclasses==0.6
decorator==5.0.9
defusedxml==0.6.0
diff-match-patch==20181111
distributed==2.11.0
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.15.2
easydict==1.9
einops==0.3.0
entrypoints==0.3
et-xmlfile==1.0.1
exceptiongroup==1.2.2
# Editable install with no version control (fairseq==0.10.2)
-e /home/nikhil/anaconda3/lib/python3.7/site-packages
faiss-cpu==1.7.4
fastapi==0.103.2
fastcache==1.1.0
filelock==3.0.12
fire==0.4.0
flake8==3.7.9
Flask==1.1.4
Flask-Cors==3.0.10
flask-paginate==0.8.1
fsspec==0.6.2
future==0.18.2
gensim==4.2.0
gevent==1.4.0
gitdb==4.0.9
GitPython==3.1.27
glob2==0.7
gmpy2==2.0.8
google==3.0.0
google-api-core==1.31.2
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-cloud==0.34.0
google-cloud-core==1.7.2
google-cloud-translate==1.7.0
google-cloud-vision==2.4.2
googleapis-common-protos==1.53.0
googletrans==3.0.0
greenlet==0.4.15
grpcio==1.39.0
grpcio-status==1.44.0
h11==0.9.0
h2==3.2.0
h5py==2.10.0
HeapDict==1.0.1
hpack==3.0.0
hstspreload==2024.11.1
html5lib==1.0.1
httpcore==0.9.1
httptools==0.6.0
httpx==0.13.3
huggingface-hub==0.0.12
hydra-core==1.0.7
hypercorn==0.15.0
hyperframe==5.2.0
hypothesis==5.5.4
idna==2.8
imageio==2.6.1
imagesize==1.2.0
importlib-metadata==1.5.0
importlib-resources==5.2.2
intervaltree==3.0.2
ipadic==1.0.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.1
jeepney==0.4.2
Jinja2==2.11.3
jmespath==0.10.0
joblib==0.14.1
json5==0.9.1
jsonlines==3.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
keyring==21.1.0
kiwisolver==1.1.0
langcodes==3.3.0
lazy-loader==0.4
lazy-object-proxy==1.4.3
libarchive-c==2.8
librosa==0.10.2.post1
lief==0.9.0
llvmlite==0.39.1
lmdb==1.3.0
locket==0.2.0
lxml==4.5.0
MarkupSafe==2.0.1
-e git+https://github.com/mantra-inc/maskrcnn_benchmark_manga@537f926ba2880c1e475c4ad9e8fada400ed188bc#egg=maskrcnn_benchmark&subdirectory=../../../third_party/maskrcnn_benchmark_manga
matplotlib==3.1.3
mccabe==0.6.1
mecab-python3==1.0.3
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.1
more-itertools==8.2.0
mpmath==1.1.0
msgpack==1.0.5
multipledispatch==0.6.0
murmurhash==1.0.6
natsort==8.1.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
ninja==1.10.2.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.56.4
numexpr==2.7.1
numpy==1.21.2
numpydoc==0.9.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.1
olefile==0.46
omegaconf==2.0.6
opencv-python==4.5.5.62
openpyxl==3.0.3
packaging==21.0
pandas==1.0.1
pandocfilters==1.4.2
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
pathy==0.6.1
patsy==0.5.1
peft==0.3.0
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
platformdirs==4.0.0
pluggy==0.13.1
ply==3.11
pooch==1.8.2
portalocker==2.4.0
preshed==3.0.6
priority==2.0.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.3
proto-plus==1.19.0
protobuf==3.20.0
psd-tools==1.9.18
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.5
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.5
pydantic==1.8.2
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.15.7
PySocks==1.7.1
pyspellchecker==0.8.1
pytest==5.3.5
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
python-magic==0.4.27
python-multipart==0.0.8
pytz==2021.1
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rope==0.16.0
rsa==4.7.2
Rtree==0.9.3
ruamel-yaml==0.15.87
s3transfer==0.5.0
sacrebleu==2.0.0
sacremoses==0.0.46
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
SecretStorage==3.1.2
Send2Trash==1.5.0
sentencepiece==0.1.96
sentry-sdk==1.9.8
seqeval==1.2.2
setproctitle==1.3.2
shortuuid==1.0.9
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.3.1
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soundfile==0.12.1
soupsieve==1.9.5
soxr==0.3.7
spacy==3.2.2
spacy-legacy==3.0.9
spacy-loggers==1.0.1
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.2.0
spyder==4.0.1
spyder-kernels==1.8.1
SQLAlchemy==1.3.13
srsly==2.4.2
starlette==0.27.0
statsmodels==0.11.0
sympy==1.5.1
tables==3.6.1
tabulate==0.8.9
taskgroup==0.0.0a4
tblib==1.6.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textaugment==2.0.0
textblob==0.17.1
thinc==8.0.13
tokenizers==0.10.3
tomli==2.0.1
toolz==0.10.0
torch==1.13.1
torchaudio==0.13.1
torchdata==0.5.1
torchtext==0.14.1
torchvision==0.14.1
tornado==6.0.3
tqdm==4.62.1
traitlets==4.3.3
transformers==4.9.2
trimesh==4.4.1
truecase==0.0.14
typer==0.4.0
typing-extensions==4.7.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.26.12
uvicorn==0.22.0
uvloop==0.18.0
wandb==0.13.3
wasabi==0.9.0
watchdog==0.10.2
watchfiles==0.20.0
wcwidth==0.1.8
webencodings==0.5.1
websockets==11.0.3
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
wsproto==1.2.0
wurlitzer==2.0.0
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
xmltodict==0.12.0
yacs==0.1.8
yapf==0.28.0
zict==1.0.0
zipp==3.7.0
